Cleanup GameController barcode parser implementation
